[FIX+IMP] spreadsheet_dashboard_oca: Define data name + error on dashboard_sheet + dynamic + new dashboard
- Add field to define name of table and restructure order of fields

- Error with dashboard_sheet selected

The error is happening because the dashboard where
the data has to be added is not being showed.

With this changes the field to define the
dashboard is required and is showed.

Added a way to limit the search to the dashboards
that can be edited

- Allow to generate dynamic tables

On dashboards exists tables that are updated dynamically,
with this changes we have the option to generate this
dynamic tables using the pivot inserter.

Added a button to duplicate and change to dynamic
table the selected table on spreadsheet editor.

If a table with multiple row goups generated, the
way to generate the dynamic table will be set the
number of rows for each level of indentation.

Example:
2 rowGroupBys
3 number of rows
- val1
    - subval1.1
    - subval1.2
    - subval1.3
- val2
    - subval2.1
    - subval2.2
    - subval2.3
- val3
    - subval3.1
    - subval3.2
    - subval3.3

- Add option to create new dashboard