[IMP] spreadsheet_*_oca: Add possibility to make columns dynamic
With these changes, we have added the ability to make the columns
dynamic, just as is done with the rows.

To add this type of table, we will need to select the dynamic columns
option and choose the number of columns to insert. Keep in mind that if
multiple column levels are selected, the chosen number will apply to
each of these levels.

In addition to this, the wizard that opens with the spreadsheet button
has been updated to allow us to choose the type of dynamism we expect in
the inserted table.