[FIX] spreadsheet_oca: Show the "Add" button in invoice list
If any module males a primary template from the ones where the button is
added, the button wasn't be shown previously.

That's because the inheritance and load order works in OWL. The way to
fix this is to make this load order deterministic specifying to be
loaded after the corresponding template for being taken into account for
primary views. This also required to split the templates file into
several ones.