[FIX] spreadsheet_oca: Clean searchParams of problematic context keys
Before these changes, when a table or chart was added with an active
user filter, the user's selections were not respected, either in the
groupings or in the measures.

Thanks to these changes, we remove the conflicting keys from the context
passed through the searchParams, thus respecting the user's selection in
the table/chart itself.

Steps to reproduce the problem:

1. Go to a pivot or graph
2. Save a filter
3. Modify the groups or measures
4. Add to a spreadsheet

Current behavior: The table will not be added with the current selection
but it will be added with the filters groups and measures.

Expected behavior:  The table will be added with the current selection.